+++ assets/javascripts/kitten/components/embed/embed-player/index.js
@@ -7,11 +7,7 @@ import { ScreenConfig } from '../../../constants/screen-config'
 import { pxToRem } from '../../../helpers/utils/typography'
 import classNames from 'classnames'
 import styled from 'styled-components'
-
-const actionKeys = ['Enter', ' ']
-
-const playerButtonSize = 90
-const playerButtonXSSize = 70
+import { PlayerIconNext } from '../../../components/graphics/icons-next/player-icon-next'
 
 const StyledEmbedPlayer = styled.div`
   position: relative;
@@ -37,63 +33,37 @@ const StyledEmbedPlayer = styled.div`
     outline: auto;
   }
 
-  &:hover .k-EmbedPlayer__button,
-  &:focus .k-EmbedPlayer__button {
-    background-color: ${COLORS.primary2};
-
-    .k-EmbedPlayer__buttonPicto {
-      fill: ${COLORS.background1};
-    }
-  }
-  &:active .k-EmbedPlayer__button {
-    background-color: ${COLORS.primary3};
-
-    .k-EmbedPlayer__buttonPicto {
-      fill: ${COLORS.background1};
-    }
-  }
-
   .k-EmbedPlayer__button {
-    width: ${pxToRem(playerButtonXSSize)};
-    height: ${pxToRem(playerButtonXSSize)};
-    top: calc(50% - ${pxToRem(playerButtonXSSize / 2)});
-    left: calc(50% - ${pxToRem(playerButtonXSSize / 2)});
-    background: ${COLORS.background1};
+    cursor: pointer;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    width: 100%;
+    height: 100%;
     position: absolute;
+    transition: opacity var(--transition), z-index var(--transition);
+    transition-delay: 0s, var(--transition-timing);
+    opacity: 1;
+    z-index: 1;
+
     display: flex;
     align-items: center;
     justify-content: center;
-    z-index: 1;
-
-    @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
-      width: ${pxToRem(playerButtonSize)};
-      height: ${pxToRem(playerButtonSize)};
-      top: calc(50% - ${pxToRem(playerButtonSize / 2)});
-      left: calc(50% - ${pxToRem(playerButtonSize / 2)});
-    }
-  }
-
-  .k-EmbedPlayer__buttonPicto {
-    width: ${pxToRem(8)};
-    height: ${pxToRem(8)};
-    @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
-      width: ${pxToRem(10)};
-      height: ${pxToRem(10)};
-    }
   }
 
   .k-EmbedPlayer__playerPreview {
     position: relative;
-    transition: opacity ease 600ms;
+    transition: opacity var(--transition), z-index var(--transition);
+    transition-delay: 0s, var(--transition-timing);
     z-index: 1;
     opacity: 1;
   }
+
+  &.k-EmbedPlayer--videoIsPlaying .k-EmbedPlayer__button,
   &.k-EmbedPlayer--videoIsPlaying .k-EmbedPlayer__playerPreview {
     opacity: 0;
     z-index: 0;
-  }
-  &.k-EmbedPlayer--cursorPointer .k-EmbedPlayer__playerPreview {
-    cursor: pointer;
   }
 `
 
@@ -102,7 +72,6 @@ export const EmbedPlayer = ({
   previewProps: { thumbnail, badgeComponent },
   iframeHtml,
   playButtonLabel,
-  style: mainStyle = void 0,
   className,
   ...others
 }) => {
@@ -116,48 +85,25 @@ export const EmbedPlayer = ({
     previewVideo.current.blur()
   }
 
-  const handleKeyPress = event => {
-    event.preventDefault()
-
-    if (actionKeys.includes(event.key)) handleClick()
-  }
-
-  const handleFocus = event => {
-    event.preventDefault()
-    previewVideo.current.focus()
-    handleKeyPress(event)
-  }
-
-  const componentClassNames = classNames('k-EmbedPlayer', className, {
-    'k-EmbedPlayer--videoIsPlaying': hasIframeHtml && isPlayerVisible,
-    'k-EmbedPlayer--cursorPointer': hasIframeHtml,
-  })
-
   return (
     <StyledEmbedPlayer
       ref={previewVideo}
       {...others}
-      style={{ ...mainStyle }}
-      onClick={hasIframeHtml ? handleClick : null}
-      onKeyPress={hasIframeHtml ? handleKeyPress : null}
-      onFocus={hasIframeHtml ? handleFocus : null}
-      role={hasIframeHtml ? 'button' : null}
-      tabIndex={hasIframeHtml ? 0 : null}
-      aria-label={hasIframeHtml ? playButtonLabel : null}
-      className={componentClassNames}
+      className={classNames('k-EmbedPlayer', className, {
+        'k-EmbedPlayer--videoIsPlaying': hasIframeHtml && isPlayerVisible,
+        'k-EmbedPlayer--cursorPointer': hasIframeHtml,
+      })}
     >
       <div className="k-EmbedPlayer__playerPreview">
         {hasIframeHtml && (
-          <div className="k-EmbedPlayer__button">
-            <svg
-              aria-hidden
-              className="k-EmbedPlayer__buttonPicto"
-              viewBox="0 0 10 10"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path d="M0 0l10 5-10 5z" />
-            </svg>
-          </div>
+          <button
+            type="button"
+            className="k-u-reset-button k-EmbedPlayer__button"
+            onClick={handleClick}
+            aria-label={playButtonLabel}
+          >
+            <PlayerIconNext />
+          </button>
         )}
 
         <ResponsiveIframeContainer ratio={validRatio}>
@@ -186,7 +132,6 @@ EmbedPlayer.propTypes = {
     thumbnail: PropTypes.shape({
       src: PropTypes.string.isRequired,
       alt: PropTypes.string.isRequired,
-      style: PropTypes.string,
     }).isRequired,
   }).isRequired,
   badgeComponent: PropTypes.node,
@@ -197,9 +142,7 @@ EmbedPlayer.propTypes = {
 
 EmbedPlayer.defaultProps = {
   previewProps: {
-    thumbnail: {
-      style: {},
-    },
+    thumbnail: {},
   },
   badgeComponent: null,
   iframeHtml: null,