+++ assets/javascripts/kitten/components/embed/video/index.js
@@ -7,12 +7,8 @@ import {
   getReactElementsWithoutType,
 } from '../../../helpers/react/get-react-elements'
 import { ScreenConfig } from '../../../constants/screen-config'
+import { PlayerIconNext } from '../../../components/graphics/icons-next/player-icon-next'
 import classNames from 'classnames'
-
-const actionKeys = ['Enter', ' ']
-
-const playerButtonSize = 90
-const playerButtonXSSize = 70
 
 const StyledVideo = styled.div`
   position: relative;
@@ -20,55 +16,28 @@ const StyledVideo = styled.div`
   height: 100%;
   overflow: hidden;
 
-  &.k-Video--hasPointerCursor {
+  .k-Video__button {
     cursor: pointer;
-  }
-
-  .k-Video__buttonContainer {
     top: 0;
+    left: 0;
     right: 0;
     bottom: 0;
-    left: 0;
+    width: 100%;
+    height: 100%;
     position: absolute;
-    transition: opacity ease 600ms, z-index ease 600ms;
-    transition-delay: 0s, 0s;
+    transition: opacity var(--transition), z-index var(--transition);
     opacity: 1;
     z-index: 1;
-    transition-delay: 0s, 600ms;
-  }
+    transition-delay: 0s, var(--transition-timing);
 
-  &.k-Video--videoIsPlaying .k-Video__buttonContainer {
-    opacity: 0;
-    z-index: 0;
-  }
-
-  .k-Video__button {
-    position: relative;
-    width: ${playerButtonXSSize};
-    height: ${playerButtonXSSize};
-    top: calc(50% - ${playerButtonXSSize} / 2);
-    left: calc(50% - ${playerButtonXSSize} / 2);
-    background: ${COLORS.background1};
     display: flex;
     align-items: center;
     justify-content: center;
-
-    @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
-      width: ${pxToRem(playerButtonSize)};
-      height: ${pxToRem(playerButtonSize)};
-      top: calc(50% - ${pxToRem(playerButtonSize / 2)});
-      left: calc(50% - ${pxToRem(playerButtonSize / 2)});
-    }
   }
 
-  .k-Video__buttonPicto {
-    width: ${pxToRem(8)};
-    height: ${pxToRem(8)};
-
-    @media (min-width: ${pxToRem(ScreenConfig.S.min)}) {
-      width: ${pxToRem(10)};
-      height: ${pxToRem(10)};
-    }
+  &.k-Video--videoIsPlaying .k-Video__button {
+    opacity: 0;
+    z-index: 0;
   }
 
   .k-Video__videoElement {
@@ -119,43 +88,24 @@ export const Video = ({
     setPlayerVisibility(!isPlayerVisible)
   }
 
-  const handleKeyDown = event => {
-    if (actionKeys.includes(event.key)) {
-      event.preventDefault()
-      handlePlayClick()
-    }
-  }
-
-  const componentClassNames = classNames('k-Video', className, {
-    'k-Video--hasPointerCursor': !controls || !autoPlay,
-    'k-Video--objectFitCover': !controls,
-    'k-Video--videoIsPlaying': isVideoPlaying,
-  })
-
   return (
     <StyledVideo
-      onClick={handlePlayClick}
-      aria-label={!autoPlay ? ariaLabel : null}
-      onKeyDown={!autoPlay ? handleKeyDown : null}
-      role={!autoPlay ? 'button' : null}
-      tabIndex={!autoPlay ? 0 : null}
-      className={componentClassNames}
+      className={classNames('k-Video', className, {
+        'k-Video--objectFitCover': !controls,
+        'k-Video--videoIsPlaying': isVideoPlaying,
+      })}
     >
       {loader}
 
       {!autoPlay && (
-        <div className="k-Video__buttonContainer">
-          <div className="k-Video__button">
-            <svg
-              aria-hidden
-              className="k-Video__buttonPicto"
-              viewBox="0 0 10 10"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path d="M0 0l10 5-10 5z" />
-            </svg>
-          </div>
-        </div>
+        <button
+          onClick={handlePlayClick}
+          aria-label={ariaLabel}
+          type="button"
+          className="k-u-reset-button k-Video__button"
+        >
+          <PlayerIconNext />
+        </button>
       )}
 
       <video